Update argument parsing to occur before index initialization
